API/requirements.txt
Flask
Werkzeug
psycopg2-binary
